UPDATE : 코스넘버 별로 닉네임빼고 묶음처리해서 가져오기 ex) courseno 가 odiga_2 면 2인 항목들 묶음 해서 가져옴 , 이미지 누르면 좌표갱신
